Articuno Bot Backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
httpx
aioredis
python-dotenv
pydantic
